Rust - modify private method and unit test
Make the private method `linked_titles_in_html` into a public associated
function. `self` is not used, and it acts as a related utility function.

Move the function's unit test into the shared `tests` directory with the
rest of the rust crawler tests.
