+++ tests/test_crawler.rs
@@ -1,3 +1,5 @@
+use std::collections::HashSet;
+
 use crawler_rs::{KEVIN_BACON_TITLE, WikipediaCrawler};
 
 const FOOTLOOSE_TITLE: &str = "Footloose_(1984_film)";
@@ -74,4 +76,103 @@ async fn three_hops() {
     assert_eq!(result.len(), 4);
     assert_eq!(result.first().unwrap(), THE_BET_TITLE);
     assert_eq!(result.last().unwrap(), KEVIN_BACON_TITLE);
+}
+
+#[tokio::test]
+async fn get_links_in_sample_html() {
+    let html = r#"
+      <p id="mwEA">
+        He is known for directing musical and comedies such as
+        <i id="mwEQ">
+            <a href="./Goodbye,_Mr._Chips_(1969_film)" id="mwEg" rel="mw:WikiLink" title="Goodbye, Mr. Chips (1969 film)">
+            Goodbye, Mr. Chips
+            </a>
+        </i>
+        (1969),
+        <i id="mwEw">
+            <a href="./The_Owl_and_the_Pussycat_(film)" id="mwFA" rel="mw:WikiLink" title="The Owl and the Pussycat (film)">
+            The Owl and the Pussycat
+            </a>
+        </i>
+        (1970),
+        <i id="mwFQ">
+            <a href="./Play_It_Again,_Sam_(film)" id="mwFg" rel="mw:WikiLink" title="Play It Again, Sam (film)">
+            Play It Again, Sam
+            </a>
+        </i>
+        (1972),
+        <i id="mwFw">
+            <a href="./The_Sunshine_Boys_(1975_film)" id="mwGA" rel="mw:WikiLink" title="The Sunshine Boys (1975 film)">
+            The Sunshine Boys
+            </a>
+        </i>
+        ,
+        <i id="mwGQ">
+            <a href="./Funny_Lady" id="mwGg" rel="mw:WikiLink" title="Funny Lady">
+            Funny Lady
+            </a>
+        </i>
+        (both 1975),
+        <i id="mwGw">
+            <a href="./The_Goodbye_Girl" id="mwHA" rel="mw:WikiLink" title="The Goodbye Girl">
+            The Goodbye Girl
+            </a>
+        </i>
+        (1977),
+        <i id="mwHQ">
+            <a href="./California_Suite_(film)" id="mwHg" rel="mw:WikiLink" title="California Suite (film)">
+            California Suite
+            </a>
+        </i>
+        (1978), and
+        <i id="mwHw">
+            <a href="./Pennies_from_Heaven_(1981_film)" id="mwIA" rel="mw:WikiLink" title="Pennies from Heaven (1981 film)">
+            Pennies From Heaven
+            </a>
+        </i>
+        (1981). His later films include
+        <i id="mwIQ">
+            <a href="./Footloose_(1984_film)" id="mwIg" rel="mw:WikiLink" title="Footloose (1984 film)">
+            Footloose
+            </a>
+        </i>
+        (1984), and
+        <i id="mwIw">
+            <a href="./Steel_Magnolias" id="mwJA" rel="mw:WikiLink" title="Steel Magnolias">
+            Steel Magnolias
+            </a>
+        </i>
+        (1989). For the drama
+        <i id="mwJQ">
+            <a href="./The_Turning_Point_(1977_film)" id="mwJg" rel="mw:WikiLink" title="The Turning Point (1977 film)">
+            The Turning Point
+            </a>
+        </i>
+        (1977) he received two
+        <a class="mw-redirect" href="./Academy_Award" id="mwJw" rel="mw:WikiLink" title="Academy Award">
+            Academy Award
+        </a>
+        .
+      </p>"#;
+    assert_eq!(
+        WikipediaCrawler::linked_titles_in_html(html),
+        HashSet::from_iter(
+            [
+                "Goodbye,_Mr._Chips_(1969_film)",
+                "The_Owl_and_the_Pussycat_(film)",
+                "Play_It_Again,_Sam_(film)",
+                "The_Sunshine_Boys_(1975_film)",
+                "Funny_Lady",
+                "The_Goodbye_Girl",
+                "California_Suite_(film)",
+                "Pennies_from_Heaven_(1981_film)",
+                "Footloose_(1984_film)",
+                "Steel_Magnolias",
+                "The_Turning_Point_(1977_film)",
+                "Academy_Award",
+            ]
+            .map(String::from)
+            .into_iter()
+        )
+    );
 }
